+++ tests/bot/presenceRouting.test.ts
@@ -198,6 +198,7 @@ describe("presence routing", () => {
     ["dev_reset_monster_rest", {}],
     ["dev_add_bandage", {}],
     ["dev_reset_yeger_bandage", {}],
+    ["dev_reset_yeger_bandage_day", {}],
     ["dev_help", {}]
   ])("routes command /%s", (command, expected) => {
     expect(getCommandPresenceContext(command)).toEqual(expected);